### 2022-11-25 #4963 @editorialbot
[PRE REVIEW]: flowMC: Normalizing-flow enhanced sampling package for probabilistic inference in Jax
**Submitting author:** @kazewong (<a href="http://orcid.org/0000-0001-8432-7788">Kaze W. K. Wong</a>)
**Repository:** https://github.com/kazewong/flowMC
**Branch with paper.md** (empty if default branch): 40-joss-paper
**Version:** v0.0.7
