### 2022-11-29 #4963 @rkurchin
👋 @GregoryAshton and @matt-graham, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
