### 2022-12-08 6585fc2
HMC now output Hamiltonian instead of logL. Add a sign to flip it.
