### 2022-12-16 #5021 @rkurchin
@kazewong do fix that one missing DOI when you have a moment

Reviewers @matt-graham and @Daniel-Dodd, let me know if you have any questions about getting your reviews started!
