### 2022-12-16 #5021 @xuanxu
The problem was that the reviewer [added initially](https://github.com/openjournals/joss-reviews/issues/4963#issuecomment-1355130615) in the pre-review issue was `@daniel-dodd` instead of `@Daniel-Dodd`, and for the bot those are two different usernames (GitHub shows both cases as the canonical form so it's difficult to see the difference unless you edit the comment)
