### 2023-01-04 kazewong/flowMC#74 @matt-graham
Installation fails in Python 3.11 environment due to TensorFlow dependency
Raising issue as part of JOSS review openjournals/joss-reviews/issues/5021

The `README.md` file lists the Python version requirements as Python 3.8+ and the `setup.cfg` file has `python_requires =  >=3.7` however on trying to install `flowmc` using `pip` into a Python 3.11 environment, installation fails with a `ERROR: Failed building wheel for tensorstore` error. This appears to be due to wheels only being available for TensorFlow (and related dependencies) for Python versions 3.7 to 3.10 currently. Given Python 3.11 is now in stable release it would be helpful to put more specific Python version requirements in the `README.md` and `setup.cfg` files. It would also make sense to have the minimum Python version in the `README.md` and `setup.cfg` files match (currently respectively 3.8 and 3.7) - as the latest versions of NumPy and JAX now support Python 3.8+ in line with [NEP 29](https://numpy.org/neps/nep-0029-deprecation_policy.html), a minimum of Python 3.8 might make sense.
