### 2023-01-04 kazewong/flowMC#75 @matt-graham
Undefined name `make_mala_sampler` imported from `flowMC.sampler.MALA` in `test/test_quickstart.py` 
Raising issue as part of JOSS review https://github.com/openjournals/joss-reviews/issues/5021

The import statement in

https://github.com/kazewong/flowMC/blob/6585fc2b8fea4a9575459e62b7c3ad209f005f15/test/test_quickstart.py#L4

fails as `make_mala_sampler` is not defined in `src/flowMC/sampler/MALA.py`.
