### 2023-01-04 kazewong/flowMC#76 @matt-graham
No separate requirements specified for tests
Raising issue as part of JOSS review https://github.com/openjournals/joss-reviews/issues/5021

The tests seem to have additional dependencies beyond those of the `flowMC` package - for example `test/test_normalizingFlow.py` imports from `sklearn`, `optax` and `matplotlib`, none of which are specified in the requirements in the `README.md` or `setup.cfg` files.

It would be useful to document the additional requirements for running the tests - for example by [adding an `[options.extras_require]` section to the `setup.cfg` file](https://setuptools.pypa.io/en/latest/userguide/declarative_config.html#configuring-setup-using-setup-cfg-files) with the additional test dependencies and/or listing the dependencies for running the tests in the `README.md` file.
