### 2023-01-04 kazewong/flowMC#77 @matt-graham
No documentation of how to run tests
Raising issue as part of JOSS review https://github.com/openjournals/joss-reviews/issues/5021

There are what appear to be a set of tests in the `test` directory at the root of the repository but there are no details given that I can see of how these are expected to be run - for example is it expected that `pytest` is used to run the tests?
