### 2023-01-04 kazewong/flowMC#79 @matt-graham
No tests for correctness of components
Raising issue as part of JOSS review https://github.com/openjournals/joss-reviews/issues/5021

The tests in the test modules in the `test` directory do not appear to have any assert statements checking that the components being tested behave as expected, but instead appear to be more integration tests that will only fail if there is an error in functions called. While such tests are useful, I would say it is also essential to have unit tests for correctness of the key individual components in the package that check for invariants being maintained, output from functions being of the expected types / shapes / values and so on.

A non-exhaustive list of some things that could be checked

  * The Markov kernels which make local proposals governed by a step-size parameter and with a Metropolis-Hastings accept step always accept / give an acceptance probability of 1, when the step size is zero corresponding to proposing to move to the current point. It could also be checked that the acceptance probability on average decreases monotonically as the step size is made larger.
  * The leapfrog integrator used in the Hamiltonian Monte Carlo implementation is time reversible.
  * The various Markov kernels produce Markov chains with empirical moments matching known analytic values to within some Monte Carlo error which scales inversely with the square root of the number of samples for a tractable known target distribution such as a standard normal (with the chains initialised from samples from the target distribution to avoid issues with initial convergence to stationarity)
  * The Markov kernels behave deterministically when passed random number generators with fixed initial states / seeds.
  * The normalizing flow models are able to learn a close approximation to a simple target distribution such as a multivariate normal distribution with random mean and covariance matrix.
  * The Jacobian log determinant iteratively computed in evaluating the normalizing flow model is consistent with the log determinant of the Jacobian of the forward map computed explicitly using the JAX `jacobian` differential operator.
  * The normalizing flow map (overall) and affine coupling components are bijections.
