### 2023-01-04 9bba584
Address typos in paper and references.
