+++ joss/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'flowMC: Normalizing-flow enhanced sampling package for probabilistic inference in Jax'
+title: 'flowMC: Normalizing flow enhanced sampling package for probabilistic inference in JAX'
 tags:
   - Python
   - Bayesian Inference 
@@ -30,44 +30,44 @@ bibliography: paper.bib
 
 # Summary
 
-Across scientific fields, more and more flexible models are required to understand increasingly complex physical processes. However the estimation of models'parameters becomes more challenging as the dimension of the parameter space grows. A common strategy to explore parameter space is to sample through a Markov Chain Monte Carlo (MCMC). Yet even MCMC methods can struggle to faithfully represent the parameter space when only relying on local updates.
+Across scientific fields, flexible models are required to understand complex physical processes. However, the estimation of model parameters becomes challenging as the dimension of the parameter space grows. A common strategy to explore parameter space is to sample through a Markov chain Monte Carlo (MCMC). Yet even MCMC methods can struggle to faithfully represent the parameter space when only relying on local updates.
 
-`flowMC` is a Python library for accelerated Markov Chain Monte Carlo (MCMC) leveraging deep generative modelling. It is built on top of the machine learning libraries `JAX` and `Flax`. At its core, `flowMC` uses a local sampler and a learnable global sampler in tandem to efficiently sample posterior distributions. While multiple chains of the local sampler generate samples over the region of interest in the target parameter space, the package uses these samples to train a normalizing flow model, then use it to propose global jumps across the parameter space. The `flowMC`sampler can handle non-trivial geometry, such as multimodal distributions and distributions with local correlations. 
+`flowMC` is a Python library for accelerated MCMC leveraging deep generative modelling, built on top of the machine learning libraries JAX and Flax. At its core, `flowMC` uses a local sampler and a learnable global sampler in tandem to efficiently sample posterior distributions. While multiple chains of the local sampler generate samples over the region of interest in the target parameter space, the package uses these samples to train a normalizing flow model and then uses it to propose global jumps across the parameter space. The `flowMC`sampler can handle non-trivial geometry, such as multimodal distributions and distributions with local correlations. 
 
 The key features of `flowMC` are summarized in the following list:
 
-- Since `flowMC` is built on top of `JAX`, it supports gradient-based samplers through automatic differentiation such as MALA and Hamiltonian Monte Carlo (HMC).
-- `flowMC` uses state-of-the-art normalizing flow models such as Rational-Quadratic Splines to power its global sampler. These models are very efficient in capturing important features within a relatively short training time.
+- Since `flowMC` is built on top of JAX, it supports gradient-based samplers through automatic differentiation such as the Metropolis-adjusted Langevin algorithm (MALA) and Hamiltonian Monte Carlo (HMC).
+- `flowMC` uses state-of-the-art normalizing flow models such as rational quadratic splines (RQS) to power its global sampler. These models are efficient in capturing important features within a relatively short training time.
 - Use of accelerators such as GPUs and TPUs are natively supported. The code also supports the use of multiple accelerators with SIMD parallelism.
-- By default, Just-in-time (JIT) compilations are used to further speed up the sampling process. 
+- By default, just-in-time (JIT) compilations are used to further acclerate the sampling process. 
 - We provide a simple black box interface for the users who want to use `flowMC` by its default parameters, yet provide at the same time an extensive guide explaining trade-offs while tuning the sampler parameters.
 
 The tight integration of all the above features makes `flowMC` a highly performant yet simple-to-use package for statistical inference.
 
 # Statement of need
 
[…]
