+++ joss/paper.bib
@@ -18,7 +18,7 @@
   author        = {Hoffman, Matthew D and Sountsov, Pavel and Dillon, Joshua V. and Langmore, Ian and Tran, Dustin and Vasudevan, Srinivas},
   booktitle     = {1st Symposium on Advances in Approximate Bayesian Inference, 2018 1–5},
   eprint        = {1903.03704},
-  title         = {{NeuTra-lizing Bad Geometry in Hamiltonian Monte Carlo Using Neural Transport}},
+  title         = {NeuTra-lizing Bad Geometry in {H}amiltonian {M}onte {C}arlo Using Neural Transport},
   url           = {http://arxiv.org/abs/1903.03704},
   year          = {2019}
 }
@@ -33,7 +33,7 @@
   number    = {3},
   pages     = {034515},
   publisher = {American Physical Society},
-  title     = {{Flow-based generative models for Markov chain Monte Carlo in lattice field theory}},
+  title     = {Flow-based generative models for {M}arkov chain {M}onte {C}arlo in lattice field theory},
   url       = {https://link.aps.org/doi/10.1103/PhysRevD.100.034515},
   volume    = {100},
   year      = {2019}
@@ -41,7 +41,7 @@
 @inproceedings{Gabrie2021a,
   author    = {Gabri{\'{e}}, Marylou and Rotskoff, Grant M. and Vanden-Eijnden, Eric},
   booktitle = {Invertible Neural Networks, NormalizingFlows, and Explicit Likelihood Models (ICML Workshop).},
-  title     = {{Efficient Bayesian Sampling Using Normalizing Flows to Assist Markov Chain Monte Carlo Methods}},
+  title     = {Efficient {B}ayesian Sampling Using Normalizing Flows to Assist {M}arkov Chain {M}onte {C}arlo Methods},
   url       = {https://arxiv.org/abs/2107.08001},
   year      = {2021}
 }
@@ -58,7 +58,7 @@
   pages         = {3964--3979},
   pmid          = {32396070},
   publisher     = {IEEE},
-  title         = {{Normalizing Flows: An Introduction and Review of Current Methods}},
+  title         = {Normalizing Flows: An Introduction and Review of Current Methods},
   volume        = {43},
   year          = {2021}
 }
@@ -70,7 +70,7 @@
   journal       = {Journal of Machine Learning Research},
   number        = {57},
   pages         = {1--64},
-  title         = {{Normalizing Flows for Probabilistic Modeling and Inference}},
+  title         = {Normalizing Flows for Probabilistic Modeling and Inference},
   url           = {https://jmlr.org/papers/v22/19-1028.html},
   volume        = {22},
   year          = {2021}
@@ -86,7 +86,7 @@
   mendeley-groups = {project-potential-based-learning,paper-mixedkernels,paper-neutravsflex},
   number          = {2},
   pmid            = {32168605},
-  title           = {{Asymptotically unbiased estimation of physical observables with neural samplers}},
+  title           = {Asymptotically unbiased estimation of physical observables with neural samplers},
   volume          = {101},
   year            = {2020}
 }
@@ -102,7 +102,7 @@
   number          = {Mc},
   pages           = {1--13},
   pmid            = {32575304},
-  title           = {{Boosting Monte Carlo simulations of spin glasses using autoregressive neural networks}},
+  title           = {Boosting {M}onte {C}arlo simulations of spin glasses using autoregressive neural networks},
   url             = {http://arxiv.org/abs/2002.04292},
   volume          = {101},
   year            = {2020}
@@ -117,7 +117,7 @@
   journal         = {Advances in Neural Information Processing Systems},
   mendeley-groups = {paper-aistat-ergofloancao,paper-mixedkernels},
[…]
