+++ joss/paper.bib
@@ -117,7 +117,7 @@
   journal         = {Advances in Neural Information Processing Systems},
   mendeley-groups = {paper-aistat-ergofloancao,paper-mixedkernels},
   number          = {MCMC},
-  title           = {{M}arkovian score climbing: Variational inference with KL(p||q)},
+  title           = {{M}arkovian score climbing: Variational inference with {KL}(p||q)},
   volume          = {2020-Decem},
   year            = {2020}
 }
