### 2023-01-04 kazewong/flowMC#81 @Daniel-Dodd
Automated test workflows
Raising issue as part of JOSS review https://github.com/openjournals/joss-reviews/issues/5021

**Issue:** You should set up a GitHub action to automatically run your tests for PRs and commits to master,  report code-coverage.

**Nice have:** automated workflow to build your documentation, automated workflow to push package releases to PyPI.
