### 2023-01-04 kazewong/flowMC#82 @matt-graham
Suggested changes to text of JOSS paper
Raising issue as part of JOSS review https://github.com/openjournals/joss-reviews/issues/5021

Some suggested changes to the text of the JOSS paper:

_However the estimation of models'parameters_ &rightarrow;
_However the estimation of models' parameters_
(missing space between models' and parameters)

_A common strategy to explore parameter space is to sample through a Markov Chain Monte Carlo (MCMC)._ &rightarrow;
_A common strategy to explore **a model's** parameter space is **to approximately sample the posterior distribution on the parameters with a** Markov Chain Monte Carlo (MCMC) **method**._
(Markov chain Monte Carlo is a descriptor of a class of algorithms and wouldn't usually be used as an object grammatically; MCMC methods sample from a distribution so helpful to specify what distribution this is)

_Yet even MCMC methods can struggle to faithfully represent the parameter space when only relying on local updates._ &rightarrow;
_Yet even MCMC methods can struggle to faithfully represent **the posterior distribution on** the parameter space when only relying on local updates._
(the Markov chain samples generated by an MCMC method represent a distribution on the parameter space, typically a posterior distribution in a Bayesian inference context, not the parameter space itself)

_FlowMC is a Python library for accelerated Markov Chain Monte Carlo (MCMC)_ &rightarrow;
_FlowMC is a Python library implementing accelerated MCMC [methods/algorithms]_
(no need to repeat acronym definition and again better to use MCMC as a qualifier)

_At its core, FlowMC uses a local sampler and a learnable global sampler in tandem to efficiently sample posterior distributions_ &rightarrow;
_At its core, FlowMC combines Metropolis-Hastings based Markov kernels proposing local moves and global moves (using a learned approximation to the target distribution) to efficiently sample posterior distributions with non-trivial geometry_
(to me local and global sampler are too vague and in particular _global sampler_ is a bit misleading as the moves proposed using the normalizing flow model are still accepted or rejected within a Metropolis acceptance step meaning the moves are still 'local' when the proposals are rejected)

_While multiple chains of the local sampler generate samples over the region of interest in the target parameter space, the package uses these samples to train a normalizing flow model, then use it to propose global jumps across the parameter space._ &rightarrow;
_While multiple chains using Markov kernels with local moves generate approximate samples from the target posterior distribution on the parameter space, the package uses these samples to train a normalizing flow model approximating the target distribution, and uses this approximation to propose global jumps across the parameter space within a Metropolis independence sampler._

_MALA_ &rightarrow; _Metropolis adjusted Langevin algorithm (MALA)_
_GPUs_ &rightarrow; _graphics processing units (GPUs)_
_TPUs_ &rightarrow; _tensor processing units (TPUs)_
_SIMD_ &rightarrow; _single instruction multiple data (SIMD)_
(all abbreviations should be defined on first usage)

_As soon as the dimension of  exceeds 3 or 4, it is necessary to resort to a robust sampling strategy such as a MCMC_

I would not be so specific here as actually I would say MCMC methods are not necessarily the best choice for target distributions on spaces with dimension 5 with quadrature methods, rejection sampling and importance sampling methods all being potentially viable in these sort of dimensionalities. I would say something more vague like _In high dimensions_ or _For parameters spaces with more than a few dimensions_ would be better.

_`Jax`_ &rightarrow; _`JAX`_ (capitalisation should be consistent throughout)

_GPU and TPU supports_ &rightarrow; _GPU and TPU support_

_The entire algorithm belongs to the class of adaptive MCMCs_ &rightarrow; _The entire algorithm belongs to the class of adaptive MCMC methods_

_the chains previous steps_ &rightarrow; _the chains' previous steps_

_Use of Accelerator_ &rightarrow; _Use of Accelerators_

_leverage Just-In-Time_ &rightarrow; _leverage JIT_ (abbreviation has been previously defined so should be used for subsequent usages)

_the log-posterior function_ &rightarrow; _a function to evaluate the logarithm of the (unnormalized) density of the posterior distribution of interest_
