### 2023-01-05 kazewong/flowMC#83 @Daniel-Dodd
Inconsistent and missing docstring on public methods.
Raising issue as part of JOSS review https://github.com/openjournals/joss-reviews/issues/5021

All public methods should have docstrings describing their functionality. Some are missing. E.g.,

https://github.com/kazewong/flowMC/blob/6585fc2b8fea4a9575459e62b7c3ad209f005f15/src/flowMC/sampler/HMC.py#L8-L36

**Nice have/suggestion:**  give short examples in the docstrings of how to use the object.

It would be also good to address inconsistent docstring formatting. E.g.,
https://github.com/kazewong/flowMC/blob/6585fc2b8fea4a9575459e62b7c3ad209f005f15/src/flowMC/sampler/Sampler.py#L14-L40

and
https://github.com/kazewong/flowMC/blob/6585fc2b8fea4a9575459e62b7c3ad209f005f15/src/flowMC/nfmodel/realNVP.py#L8-L26
