### 2023-01-05 kazewong/flowMC#84 @Daniel-Dodd
`progress_bar_scan` credit
Raising issue as part of JOSS review https://github.com/openjournals/joss-reviews/issues/5021

The `progress_bar_scan`, i.e.,
https://github.com/kazewong/flowMC/blob/79631274e7d06a3f71ce41a7f26322f83dfb72e1/src/flowMC/utils/progressBar.py#L7-L13

is seemingly adapted from this (excellent) blog post https://www.jeremiecoullon.com/2021/01/29/jax_progress_bar/

There is no problem in using this. But you should acknowledge credit to this in the decorator's docstring.
