### 2023-01-06 fe4ae41
Adding descritpion of the loading of a pretrained models in  the docs
