### 2023-01-08 84a8f80
adding docstrings to HMC.py
