### 2023-01-08 5e87459
Using only one formatting for docstrings
