+++ src/flowMC/nfmodel/realNVP.py
@@ -8,21 +8,16 @@ import numpy as np
 class MLP(nn.Module):
 
     """
-    Multi-layer perceptron in Flax.
+    Multi-layer perceptron in Flax. We use a gaussian kernel with a standard deviation
+    of `init_weight_scale=1e-4` by default.
 
-    Parameters
-    ----------
-    features : list of int
-        The number of features in each layer.
-    activation : callable
-        The activation function at each level
-    use_bias : bool
-        Whether to use bias in the layers.
-    init_weight_scale : float
-        The initial weight scale for the layers.
-    kernel_init : callable
-        The kernel initializer for the layers.
-        We use a gaussian kernel with a standard deviation of `init_weight_scale` by default.
+    Args:
+        features: (list of int) The number of features in each layer.
+        activation: (callable) The activation function at each level
+        use_bias: (bool) Whether to use bias in the layers.
+        init_weight_scale: (float) The initial weight scale for the layers.
+        kernel_init: (callable) The kernel initializer for the layers.
+        
     """
 
     features: Sequence[int]
@@ -51,19 +46,15 @@ class MLP(nn.Module):
 class AffineCoupling(nn.Module):
 
     """
-    Affine coupling layer. (Defined in the RealNVP paper https://arxiv.org/abs/1605.08803)
+    Affine coupling layer. 
+    (Defined in the RealNVP paper https://arxiv.org/abs/1605.08803)
     We use tanh as the default activation function.
 
-    Parameters
-    ----------
-    n_features : int
-        The number of features in the input.
-    n_hidden : int
-        The number of hidden units in the MLP.
-    mask : ndarray
-        Alternating mask for the affine coupling layer.
-    dt : float
-        Scaling factor for the affine coupling layer.
+    Args:
+        n_features: (int) The number of features in the input.
+        n_hidden: (int) The number of hidden units in the MLP.
+        mask: (ndarray) Alternating mask for the affine coupling layer.
+        dt: (float) Scaling factor for the affine coupling layer.
     """
 
     n_features: int
@@ -103,20 +94,13 @@ class RealNVP(nn.Module):
     RealNVP mode defined in the paper https://arxiv.org/abs/1605.08803.
     MLP is needed to make sure the scaling between layers are more or less the same.
 
-    Parameters
-    ----------
-    n_layer : int
-        The number of affine coupling layers.
-    n_features : int
-        The number of features in the input.
-    n_hidden : int
-        The number of hidden units in the MLP.
-    dt : float
-        Scaling factor for the affine coupling layer.
-    base_mean : ndarray
-        Mean of Gaussian base distribution
-    base_cov : ndarray
-        Covariance of Gaussian base distribution
+    Args:
+        n_layer: (int) The number of affine coupling layers.
+        n_features: (int) The number of features in the input.
+        n_hidden: (int) The number of hidden units in the MLP.
+        dt: (float) Scaling factor for the affine coupling layer.
+        base_mean: (ndarray) Mean of Gaussian base distribution
+        base_cov: (ndarray) Covariance of Gaussian base distribution
     """
 
     n_layer: int
@@ -128,7 +112,7 @@ class RealNVP(nn.Module):
         affine_coupling = []
         for i in range(self.n_layer):
             mask = np.ones(self.n_features)
-            mask[int(self.n_features / 2) :] = 0
+            mask[int(self.n_features / 2):] = 0
             if i % 2 == 0:
                 mask = 1 - mask
             mask = jnp.array(mask)
