+++ src/flowMC/nfmodel/rqSpline.py
@@ -85,18 +85,12 @@ class RQSpline(nn.Module):
     """
     Rational quadratic spline normalizing flow model using distrax.
 
-    Parameters
-    ----------
-    n_features : int
-        Number of features in the data.
-    num_layers : int
-        Number of layers in the flow.
-    num_bins : int
-        Number of bins in the spline.
-    hidden_size : Sequence[int]
-        Size of the hidden layers in the conditioner.
-    spline_range : Sequence[float]
-        Range of the spline.
+    Args:
+        n_features : (int) Number of features in the data.
+        num_layers : (int) Number of layers in the flow.
+        num_bins : (int) Number of bins in the spline.
+        hidden_size : (Sequence[int]) Size of the hidden layers in the conditioner.
+        spline_range : (Sequence[float]) Range of the spline.
     """
 
     n_features: int
