### 2023-01-08 8209a55
minor docstring touches
