### 2023-01-09 kazewong/flowMC#82 @Daniel-Dodd
Comments on the paper as part of JOSS review https://github.com/openjournals/joss-reviews/issues/5021, to supplement the (excellent) review above. **NOTE** these comments with regards to commit; d2414382fd74d51ed6d8ad9b9c28dd67e0592865, on kazewong/flowMC/40-joss-paper, which is currently the latest version of the paper.

# General typo comments (in addition to the above review):

- [ ] **Title:** Normalizing flow should not have a hyphen.
- [ ] **Capitalisation:** unless, it is the first word of a sentence, or is someones name, e.g. Markov or Bayesian, do not capitalise. For example,  *`Just-In-Time` to `just-in-time` , or* `Markov Chain Monte Carlo` to `Markov chain Monte Carlo` , or `Rational-Quadratic Splines` to `rational quadratic splines` (there is no hyphen here), … etc.
- [ ] As mentioned by the other reviewer. All acronyms need to be defined on first usage, but do not need to be repeated again. Another one not given above is `HMC` as `Hamiltonian Monte Carlo` .
- [ ] You are using NF, short for, normalising flow. This is fine, but I would recommend defining this upon the first instance of normalizing flow appears (which is the _summary_ section) and using NF thereon.
- [ ] None of your math equations are correctly punctuated.
- [ ] Remove unnecessary adjectives from the manuscript. These make sentences challenging to read, e.g. `more and more` , `increasingly` , `very` , etc.

From [](https://github.com/matt-graham)the other reviewer’ feedback in the above, I would recommend you hyphenate `Metropolis-adjusted Langevin algorithm (MALA)` , and do not capitalise “chain” in `Markov chain Monte Carlo`.

For more typos please see my PR #80. Note this PR does not address all the above (or the below), it only contains some of what I have noticed.

# Structure (mostly ignoring typos here)

_“Across scientific fields, more and more flexible models are required to understand increasingly complex physical processes.”_

- [ ] Baseless claim. Need examples with citations if you wish to use this in the paper.

_“However the estimation of models’parameters becomes more challenging as the dimension of the parameter space grows.”_

- [ ] Motivate why we need high dimensional probabilistic models to model physical processes in the first place? Why does inference become challenging in higher dimensions?

_“estimation of models’parameters”?_ This is myopic for a Bayesian approach? I would talk about the challenges of inferring the posterior distribution, rather than parameter estimation (which is only one facet of Bayesian inference).

- [ ] Perhaps mention intractability of the posterior and would motivate why we are using approximate inference to begin with?

typo: _`models’parameters`_

_“*Yet even MCMC methods can struggle to faithfully represent the parameter space when only relying on local updates.”*_

- [ ] What MCMC methods? Perhaps you mean `conventional` MCMC methods? Please say so.
- [ ] Why / in what sense do they struggle? Perhaps you mean they are inefficient / take longer time to converge in higher dimensions. Can you please reference literature that highlights this issue.
- [ ] As identified by the first reviewer, parameter space is another typo.

_“The flowMCsampler can handle non-trivial geometry, such as multimodal distributions and distributions with local correlations.”_

- [ ] typo: _`flowMCsampler`_

“It is built on top of the machine learning libraries JAX and Flax.”

- [ ] Both JAX and Flax need citations.

_“While multiple chains of the local sampler generate samples over the region of interest in the target parameter space, the package uses these samples to train a normalizing flow model, then use it to propose global jumps across the parameter space.”_

- [ ]  _“the package”_ I would be specific with `flowMC`

_“The key features of flowMC are summarized in the following list:”_

- [ ] Personally, not a fan of seeing bullet lists in a paper. This could be rewritten more cohesively.

I would try and aim for two concise paragraphs at the most in this section. Though you do state some features of normalising flow enhanced MCMC, I would tell the story in a differently to make the comparison clearer to why you would use this over conventional MCMC. I would first be specific with the misgivings  of conventional MCMC, i.e., inefficient in high dimensions, state the issues with multimodal and locally correlated distributions, and then after that introduce the normalising flow approach “on the other hand” …. Perhaps mention the target audience, to draw reader in.

Already mentioned. But it is important you punctuate all maths.

***Gradient-based sampler***

_“The computational cost of obtaining a gradient in this way is often of the same order as evaluating the target function itself, making gradient-based samplers favorable with respect to the efficiency/accuracy trade-off.”_

- [ ] This is a baseless claim. Moreover, it is irrelevant if this is generally true for most automatic differentiation software. What is of relevance here: can you guarantee this true specifically for your JAX implementation, and or for JAX in general? If yes, do justify, if not then remove. This otherwise serves as an unfounded performance claim, and I advise you remove this.

**Learned transition kernels with normalizing flow**

_”Posterior distribution of many real-world problems have non-trivial geometry such as multi-modality and local correlations, which could drastically slow down the convergence of the sampler only based on gradient information.”_

- [ ] Baseless claim. Please cite relevant literature.

- [ ] Typo: _“Variant of this idea have been...”_ to _“Variant**s** of this idea have been...”_

- [ ] _“and references therein”_. If these are important, please cite, if not remove.

- [ ] _`Sequential Monte Carlo`_ to _`sequential Monte Carlo`_

- [ ] ***`Use of Accelerator`*** to ***`Use of accelerator`***
 _”Modern accelerators such as GPUs and TPUs are designed to execute dense computation in parallel. Due to the sequential nature of MCMC, a common approach in leveraging accelerators is to run multiple chains in parallel, then combine their results to obtain the posterior distribution. However, a large portion of the computation time comes from the burn-in phase for which chain-parallelization provides no speed up.”_

- [ ] Don’t see how this is relevant. I would remove this, to just have just this:

_`flowMC` is built on top of `JAX`, so that it supports the use of GPU and TPU accelerators by default. Users can write codes in the same way as they would do on a CPU, and the library will automatically detect the available accelerators and use them at run time. Furthermore, the library leverage Just-In-Time compilations to further improve the performance of the sampler._

- [ ] Note there are a few typos in the above paragraph as identified in my PR #80.

_**Simplicity and extensibility -**_ Make it clear this concerns the target audience.

_“While we provide a high-level API for most of the users, the code is also designed to be
extensible.”_

- [ ] Who are “most of the users”? Practitioners?
- [ ] Designed to be “extensible”.  Great, but for whom? Researchers?
- [ ] Make it clear who the target audience is.

# Overall

This is shaping up nicely. You give key functionality that your library has to offer and identify the advantages of a JAX codebase. Just need to fix those typos, and support some the claims made. Look forward to reading the revised version. :)
