### 2023-01-09 kazewong/flowMC#85 @Daniel-Dodd
Contribution guide
Raising issue as part of JOSS review https://github.com/openjournals/joss-reviews/issues/5021

Contribution guide is brief and assumes everyone is confident with how to use GitHub and make PRs.

 Suggestions:
 - Outline **all** the different ways someone can contribute to the project.
 - Make it clear who the maintainers are and how to contact them if someone needs help.
 - Extend the guide to show how a user can open a PR, like PyMC's step-by-step contribution guide. https://www.pymc.io/projects/docs/en/latest/contributing/pr_tutorial.html

I also recommend adding details/ a link to the contribution guide in your documentation.
