### 2023-01-12 2e69829
Add equal sign in mala_sampler autotune
