### 2023-01-17 2e6ab7a
Delete unused progress bar files
