+++ src/flowMC/sampler/LocalSampler_Base.py
@@ -4,7 +4,6 @@ class LocalSamplerBase:
 
     def __init__(self, logpdf: Callable, jit: bool, params: dict) -> Callable:
         """
-        
         """
         self.logpdf = logpdf
         self.jit = jit
@@ -13,15 +12,17 @@ class LocalSamplerBase:
     @abstractmethod
     def make_kernel(self, return_aux = False) -> Callable:
         """
-        
+        Make the kernel of the sampler for one update
         """
 
     @abstractmethod
     def make_update(self) -> Callable:
         """
+        Make the update function for multiple steps
         """
 
     @abstractmethod
     def make_sampler(self) -> Callable:
         """
+        Make the sampler for multiple chains given initial positions
         """
