### 2023-01-18 kazewong/flowMC#74 @marylou-gabrie
Thanks Matt for noticing this issue and its origin.
Following your suggestion I updated the requirement for python to be in the range 3.8-3.10.x.
