### 2023-01-18 e993a04
Keep the potential energy (negative log likelihood) instead of the hamiltonian in the HMC sampler.
