### 2023-01-25 7db0379
starting to implement additional suggested changes by reviewers
