+++ joss/paper.md
@@ -29,10 +29,15 @@ bibliography: paper.bib
 ---
 
 # Summary
+Across scientific fields, the Bayesian framework is used to account for uncertainties in inference [ellison2004bayesian,lancaster2004introduction,von2011bayesian]. However, for models with more than a few parameters, exact inference is intractable. A frequently used strategy is to approximately sample the posterior distribution on models' parameters with a Markov Chain Monte Carlo (MCMC). Yet conventional MCMC methods relying on local updates can take a prohitive time to converge when posterior distributions have complex geometries (see e.g., [@rubinstein_simulation_2017] for a textbook).
 
+`flowMC` is a Python library implementing accelerated MCMC leveraging deep generative modelling as proposed by [@Gabrie2021a], built on top of the machine learning libraries JAX [@jax2018github] and Flax [flax2020github]. At its core, `flowMC` uses a combination of Metropolis-Hastings Markov kernels using local and global proposed moves. While multiple chains are ran using local-update Markov kernels to generate approximate samples over the region of interest in the target parameter space, these samples are used to train a NF model to approximate the samples' density. The NF is then used in the independent Metropois-Hastings kernel to propose global jumps across the parameter space. The `flowMC` sampler can handle non-trivial geometry, such as multimodal distributions and distributions with local correlations. 
 
-`flowMC` is a Python library implementing accelerated MCMC leveraging deep generative modelling. It is built on top of the machine learning libraries Jax and Flax. At its core, `flowMC` uses a local sampler and a learnable global sampler in tandem to efficiently sample posterior distributions. While multiple chains of the local sampler generate samples over the region of interest in the target parameter space, the package uses these samples to train a normalizing flow model, then use it to propose global jumps across the parameter space. The `flowMC`sampler can handle non-trivial geometry, such as multimodal distributions and distributions with local correlations. 
+The key features of `flowMC` are summarized in the following list:
 
+- Since `flowMC` is built on top of JAX, it supports gradient-based samplers through automatic differentiation such as the Metropolis-adjusted Langevin algorithm (MALA) and Hamiltonian Monte Carlo (HMC).
+- `flowMC` uses state-of-the-art NF models such as rational quadratic splines (RQS) to power its global sampler. These models are efficient in capturing important features within a relatively short training time.
+- Use of accelerators such as graphics processing unites (GPUs) and tensor processing unites (TPUs) are natively supported. The code also supports the use of multiple accelerators with SIMD parallelism.
 - By default, just-in-time (JIT) compilations are used to further acclerate the sampling process. 
 - We provide a simple black box interface for the users who want to use `flowMC` by its default parameters, yet provide at the same time an extensive guide explaining trade-offs while tuning the sampler parameters.
 
@@ -47,45 +52,38 @@ p(\theta|\mathcal{D}) = \frac{\ell(\mathcal{D}|\theta) p_0(\theta)}{Z(\mathcal{D
 $$
 
 where $\ell(\mathcal{D}|\theta)$ is the likelihood induced by the model, $p_0(\theta)$ the prior on the parameters and  $Z(\mathcal{D})$ the model evidence. 
-As soon as the dimension of $\theta$ exceeds 3 or 4, it is necessary to resort to a robust sampling strategy such as MCMC. Drastic gains in computational efficiency can be obtained by a careful selection of the MCMC transition kernel which can be assisted by machine learning methods and libraries.  
+For parameter space with more than a few dimensions, it is necessary to resort to a robust sampling strategy such as MCMC. Drastic gains in computational efficiency can be obtained by a careful selection of the MCMC transition kernel which can be assisted by machine learning methods and libraries.  
 
 ***Gradient-based sampler***
 In a high dimensional space, sampling methods which leverage gradient information of the target distribution aid by proposing new samples likely to be accepted.
[…]
