+++ joss/paper.bib
@@ -177,3 +177,55 @@
   volume   = {6},
   year     = {2018}
 }
+@software{jax2018github,
+  author = {James Bradbury and Roy Frostig and Peter Hawkins and Matthew James Johnson and Chris Leary and Dougal Maclaurin and George Necula and Adam Paszke and Jake Vander{P}las and Skye Wanderman-{M}ilne and Qiao Zhang},
+  title = {{JAX}: composable transformations of {P}ython+{N}um{P}y programs},
+  url = {http://github.com/google/jax},
+  version = {0.3.13},
+  year = {2018},
+}
+@software{flax2020github,
+  author = {Jonathan Heek and Anselm Levskaya and Avital Oliver and Marvin Ritter and Bertrand Rondepierre and Andreas Steiner and Marc van {Z}ee},
+  title = {{F}lax: A neural network library and ecosystem for {JAX}},
+  url = {http://github.com/google/flax},
+  version = {0.6.3},
+  year = {2020},
+}
+@article{ellison2004bayesian,
+  title={Bayesian inference in ecology},
+  author={Ellison, Aaron M},
+  journal={Ecology letters},
+  volume={7},
+  number={6},
+  pages={509--520},
+  year={2004},
+  publisher={Wiley Online Library}
+}
+@article{von2011bayesian,
+  title={Bayesian inference in physics},
+  author={Von Toussaint, Udo},
+  journal={Reviews of Modern Physics},
+  volume={83},
+  number={3},
+  pages={943},
+  year={2011},
+  publisher={APS}
+}
+@book{lancaster2004introduction,
+  title={An introduction to modern Bayesian econometrics},
+  author={Lancaster, Tony},
+  year={2004},
+  publisher={Blackwell Oxford}
+}
+@book{rubinstein_simulation_2017,
+	address = {Hoboken, New Jersey},
+	edition = {Third edition},
+	series = {Wiley series in probability and statistics},
+	title = {Simulation and the {Monte} {Carlo} method},
+	isbn = {978-1-118-63220-8 978-1-118-63238-3},
+	publisher = {John Wiley \& Sons, Inc},
+	author = {Rubinstein, Reuven Y. and Kroese, Dirk P.},
+	year = {2017},
+	keywords = {Textbook, Digital computer simulation, Mathematical statistics, Monte Carlo method, Sampling (Statistics)},
+	file = {Rubinstein and Kroese - 2017 - Simulation and the Monte Carlo method.pdf:/Users/marylou/Zotero/storage/ZYC69LMC/Rubinstein and Kroese - 2017 - Simulation and the Monte Carlo method.pdf:application/pdf},
+}
