### 2023-01-25 f3629bc
example ref for complicated geometry
