### 2023-01-25 kazewong/flowMC#82 @marylou-gabrie
Thanks a lot @Daniel-Dodd and @matt-graham for the valuable feedback and the spotted typos.

I iterated on your suggestions in commit:
[#1827618](https://github.com/kazewong/flowMC/commit/1827618fdd86b50c94b18b1a210e508a83d18a21)

@Daniel-Dodd we are leaning towards keeping the key features itemized as we feel it makes it easy for readers to parse them.
