+++ joss/paper.md
@@ -31,7 +31,7 @@ bibliography: paper.bib
 # Summary
 Across scientific fields, the Bayesian framework is used to account for uncertainties in inference [ellison2004bayesian,lancaster2004introduction,von2011bayesian]. However, for models with more than a few parameters, exact inference is intractable. A frequently used strategy is to approximately sample the posterior distribution on models' parameters with a Markov Chain Monte Carlo (MCMC). Yet conventional MCMC methods relying on local updates can take a prohitive time to converge when posterior distributions have complex geometries (see e.g., [@rubinstein_simulation_2017] for a textbook).
 
-`flowMC` is a Python library implementing accelerated MCMC leveraging deep generative modelling as proposed by [@Gabrie2021a], built on top of the machine learning libraries JAX [@jax2018github] and Flax [flax2020github]. At its core, `flowMC` uses a combination of Metropolis-Hastings Markov kernels using local and global proposed moves. While multiple chains are ran using local-update Markov kernels to generate approximate samples over the region of interest in the target parameter space, these samples are used to train a NF model to approximate the samples' density. The NF is then used in the independent Metropois-Hastings kernel to propose global jumps across the parameter space. The `flowMC` sampler can handle non-trivial geometry, such as multimodal distributions and distributions with local correlations. 
+`flowMC` is a Python library implementing accelerated MCMC leveraging deep generative modelling as proposed by @Gabrie2022, built on top of the machine learning libraries JAX [@jax2018github] and Flax [flax2020github]. At its core, `flowMC` uses a combination of Metropolis-Hastings Markov kernels using local and global proposed moves. While multiple chains are ran using local-update Markov kernels to generate approximate samples over the region of interest in the target parameter space, these samples are used to train a NF model to approximate the samples' density. The NF is then used in the independent Metropois-Hastings kernel to propose global jumps across the parameter space. The `flowMC` sampler can handle non-trivial geometry, such as multimodal distributions and distributions with local correlations. 
 
 The key features of `flowMC` are summarized in the following list:
 
@@ -64,7 +64,7 @@ When the posterior distribution has a non-trivial geometry, such as multiple mod
 To address this problem, `flowMC` also uses a generative model, namely a NF [@Papamakarios2019; @Kobyzev2021], that is trained to mimic the posterior distribution and used as a proposal in Metropolis-Hastings MCMC steps. Variants of this idea have been explored in the past few years [e.g., @Albergo2019; @Hoffman2019; @Parno2018].
 Despite the growing interest in these methods, few accessible implementations for practitioners exist, especially with GPU and TPU support. Notably, a version of the NeuTra sampler [@Hoffman2019] is available in Pyro [@bingham2019pyro], and the PocoMC package [@Karamanis2022] implements a version of sequential Monte Carlo (SMC), including NFs.
 
-`flowMC` implements the method proposed by @Gabrie2021a. 
+`flowMC` implements the method proposed by @Gabrie2022. 
 As individual chains explore their local neighborhood through gradient-based MCMC steps, multiple chains train the NF to learn the global landscape of the posterior distribution. In turn, the chains can be propagated with a Metropolis-Hastings kernel using the NF to propose globally in the parameter space. The cycle of local sampling, NF tuning, and global sampling is repeated until obtaining chains of the desired length.
 The entire algorithm belongs to the class of adaptive MCMC methods [@Andrieu2008], collecting information from the chains' previous steps to simultaneously improve the transition kernel. 
 Usual MCMC diagnostics can be applied to assess the robustness of the inference results, thereby avoiding the common concern of validating the NF model. 
