+++ joss/paper.bib
@@ -1,4 +1,4 @@
-@article{Gabrie2021,
+@article{Gabrie2022,
   author          = {Gabri{\'{e}}, Marylou and Rotskoff, Grant M. and Vanden-Eijnden, Eric},
   doi             = {10.1073/pnas.2109420119},
   eprint          = {2105.12603},
@@ -228,4 +228,14 @@
 	year = {2017},
 	keywords = {Textbook, Digital computer simulation, Mathematical statistics, Monte Carlo method, Sampling (Statistics)},
 	file = {Rubinstein and Kroese - 2017 - Simulation and the Monte Carlo method.pdf:/Users/marylou/Zotero/storage/ZYC69LMC/Rubinstein and Kroese - 2017 - Simulation and the Monte Carlo method.pdf:application/pdf},
+}
+@article{neal2003slice,
+  title={Slice sampling},
+  author={Neal, Radford M},
+  journal={The annals of statistics},
+  volume={31},
+  number={3},
+  pages={705--767},
+  year={2003},
+  publisher={Institute of Mathematical Statistics}
 }
