### 2023-01-25 kazewong/flowMC#82 @Daniel-Dodd
Nice work, @marylou-gabrie! Paper looks good to me - the story is nice and clear. Happy with the itemized list. :)

Just noticed a couple of things on [84c7fdf](https://github.com/kazewong/flowMC/commit/84c7fdf554c724562f931c02bda3ec8f01edcbd7). Think, @matt-graham, left a few comments on your commit changes, but I'll give mine here. I believe that some of your references aren't appearing on the [compiled pdf](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05021/joss.05021/10.21105.joss.05021.pdf), perhaps due to missing an `@` symbol? Also, I spotted a few remaining typos:

## Intro
Paragraph 1:
- [ ] _prohitive_ -> _prohibitive_

Paragraph 2:
- [ ] _While multiple chains are **ran** using local-update Markov kernels to ...._ -> _While multiple chains are **run** using local-update Markov kernels to ..._

Itemised list:
- [ ] _Use of accelerators such as graphics processing **unites** (GPUs) and tensor processing **unites** (TPUs) are natively supported. The code also supports the use of multiple accelerators with SIMD parallelism._ -> _Use of accelerators such as graphics processing *units* (GPUs) and tensor processing *units* (TPUs) are natively supported. The code also supports the use of multiple accelerators with SIMD parallelism._
- [ ] _By default, just-in-time (JIT) compilations are used to further **acclerate** the sampling process._ -> _By default, just-in-time (JIT) compilations are used to further **accelerate** the sampling process._

## Statement of need:
- [ ] _The mathematical **detail** of the method are explained in [@Gabrie2021a]._ -> _The mathematical **details** of the method are explained in [@Gabrie2021a]._

- [ ] _While we provide a high-level API suitable for most **practitionners**, the code is also designed to be extensible._ -> _While we provide a high-level API suitable for most **practitioners**, the code is also designed to be extensible._
