### 2023-01-28 9bfc5d6
Add question about code to contributing guide
