### 2023-02-22 d248815
Deterministic kernel test passing
