+++ test/unit/test_kernels.py
@@ -25,28 +25,50 @@ def test_HMC_deterministic():
     result1 = (HMC_kernel(rng_key_set[0], initial_position[0], initial_PE[0], HMC_obj.params))
     result2 = (HMC_kernel(rng_key_set[0], initial_position[0], initial_PE[0], HMC_obj.params))
 
-    assert result1==result2
+    assert (result1[0]==result2[0]).all()
+    assert result1[1]==result2[1]
+    assert result1[2]==result2[2]
 
 def test_MALA_deterministic():
     n_dim = 2
     n_chains = 1
-    MALA_obj = MALA(log_posterior, True, {"step_size": 1, "inverse_metric": jnp.ones(n_dim)})
+    MALA_obj = MALA(log_posterior, True, {"step_size": 1})
     MALA_kernel = MALA_obj.make_kernel()
 
     rng_key_set = initialize_rng_keys(n_chains, seed=42)
 
     initial_position = jax.random.normal(rng_key_set[0], shape=(n_chains, n_dim)) * 1
 
-    initial_PE = jax.vmap(MALA_obj.potential)(initial_position)
+    initial_logp = log_posterior(initial_position)
 
     MALA_kernel = MALA_obj.make_kernel()
 
-    result1 = (MALA_kernel(rng_key_set[0], initial_position[0], initial_PE[0], MALA_obj.params))
-    result2 = (MALA_kernel(rng_key_set[0], initial_position[0], initial_PE[0], MALA_obj.params))
+    result1 = (MALA_kernel(rng_key_set[0], initial_position[0], initial_logp, MALA_obj.params))
+    result2 = (MALA_kernel(rng_key_set[0], initial_position[0], initial_logp, MALA_obj.params))
 
-    assert result1==result2
+    assert (result1[0]==result2[0]).all()
+    assert result1[1]==result2[1]
+    assert result1[2]==result2[2]
 
 def test_Gaussian_random_walk_deterministic():
-    assert 1
+    n_dim = 2
+    n_chains = 1
+    GRW_obj = GaussianRandomWalk(log_posterior, True, {"step_size": 1})
+    GRW_kernel = GRW_obj.make_kernel()
+
+    rng_key_set = initialize_rng_keys(n_chains, seed=42)
+
+    initial_position = jax.random.normal(rng_key_set[0], shape=(n_chains, n_dim)) * 1
+
+    initial_logp = log_posterior(initial_position)
+
+    GRW_kernel = GRW_obj.make_kernel()
+
+    result1 = (GRW_kernel(rng_key_set[0], initial_position[0], initial_logp, GRW_obj.params))
+    result2 = (GRW_kernel(rng_key_set[0], initial_position[0], initial_logp, GRW_obj.params))
+
+    assert (result1[0]==result2[0]).all()
+    assert result1[1]==result2[1]
+    assert result1[2]==result2[2]
 
 
