### 2023-02-23 4d8e8d9
wip tests for realnvps
