### 2023-02-23 8b3198a
Can train a rqSpline in test_nf.py now
