+++ test/unit/test_nf.py
@@ -27,32 +27,30 @@ def test_rqSpline():
     )
 
     # Model parameters
-    n_layers = 8
-    n_hiddens = [64, 64]
+    n_layers = 2
+    n_hiddens = [16, 16]
     n_bins = 8
-    n_dim = 2
 
     model = RQSpline(n_dim, n_layers, n_hiddens, n_bins)
 
-    variables = model.init(rng_model, jnp.ones((1, 2)))["variables"]
+    variables = model.init(rng_model, jnp.ones((1, n_dim)))["variables"]
     variables = variables.unfreeze()
     variables["base_mean"] = jnp.mean(data, axis=0)
     variables["base_cov"] = jnp.cov(data.T)
     variables = flax.core.freeze(variables)
 
     def create_train_state(rng, learning_rate, momentum):
-        params = model.init(rng, jnp.ones((1, 2)))["params"]
+        params = model.init(rng, jnp.ones((1, n_dim)))["params"]
         tx = optax.adam(learning_rate, momentum)
         return train_state.TrainState.create(apply_fn=model.apply, params=params, tx=tx)
-
-    state = create_train_state(rng_init, learning_rate, momentum)
 
     # Optimization parameters
     num_epochs = 3
     batch_size = 10000
     learning_rate = 0.001
     momentum = 0.9
-
+    state = create_train_state(rng_init, learning_rate, momentum)
+    
     train_flow, train_epoch, train_step = make_training_loop(model)
     rng, state, loss_values = train_flow(
         rng_train, state, variables, data, num_epochs, batch_size
