+++ test/unit/test_kernels.py
@@ -8,10 +8,10 @@ import jax.numpy as jnp
 def log_posterior(x):
     return -0.5 * jnp.sum(x ** 2)
 
-def test_HMC_deterministic():
+def test_HMC():
     n_dim = 2
     n_chains = 1
-    HMC_obj = HMC(log_posterior, True, {"step_size": 1,"n_leapfrog": 3, "inverse_metric": jnp.ones(n_dim)})
+    HMC_obj = HMC(log_posterior, True, {"step_size": 1,"n_leapfrog": 5, "inverse_metric": jnp.ones(n_dim)})
     HMC_kernel = HMC_obj.make_kernel()
 
     rng_key_set = initialize_rng_keys(n_chains, seed=42)
@@ -20,14 +20,31 @@ def test_HMC_deterministic():
 
     initial_PE = jax.vmap(HMC_obj.potential)(initial_position)
 
-    HMC_kernel = HMC_obj.make_kernel()
+    HMC_kernel, leapfrog_kernel, leapfrog_step = HMC_obj.make_kernel(return_aux=True)
+
+    # Test whether the HMC kernel is deterministic
 
     result1 = (HMC_kernel(rng_key_set[0], initial_position[0], initial_PE[0], HMC_obj.params))
     result2 = (HMC_kernel(rng_key_set[0], initial_position[0], initial_PE[0], HMC_obj.params))
 
-    assert (result1[0]==result2[0]).all()
+    assert jnp.allclose(result1[0],result2[0])
     assert result1[1]==result2[1]
     assert result1[2]==result2[2]
+
+    # Test whether the leapfrog kernel is reversible
+
+    key1, key2 = jax.random.split(rng_key_set[0])
+    initial_momentum = jax.random.normal(key1, shape=initial_position.shape) * jnp.ones(n_dim)**-0.5
+
+    new_position, new_momentum = leapfrog_step(initial_position, initial_momentum, HMC_obj.params)
+    rev_position, rev_momentum = leapfrog_step(new_position, -new_momentum, HMC_obj.params)
+
+    assert jnp.allclose(rev_position, initial_position)
+    assert jnp.allclose(initial_PE, HMC_obj.potential(rev_position))
+
+    # Test whether the HMC kernel can recover moment of a Gaussian
+
+    # Test acceptance rate goes to one when step size is small
 
 def test_MALA_deterministic():
     n_dim = 2
@@ -46,7 +63,7 @@ def test_MALA_deterministic():
     result1 = (MALA_kernel(rng_key_set[0], initial_position[0], initial_logp, MALA_obj.params))
     result2 = (MALA_kernel(rng_key_set[0], initial_position[0], initial_logp, MALA_obj.params))
 
-    assert (result1[0]==result2[0]).all()
+    assert jnp.allclose(result1[0],result2[0])
     assert result1[1]==result2[1]
     assert result1[2]==result2[2]
 
@@ -67,8 +84,16 @@ def test_Gaussian_random_walk_deterministic():
     result1 = (GRW_kernel(rng_key_set[0], initial_position[0], initial_logp, GRW_obj.params))
     result2 = (GRW_kernel(rng_key_set[0], initial_position[0], initial_logp, GRW_obj.params))
 
-    assert (result1[0]==result2[0]).all()
+    assert jnp.allclose(result1[0],result2[0])
     assert result1[1]==result2[1]
     assert result1[2]==result2[2]
 
+def test_HMC_moment():
+    pass
+
+def test_MALA_moment():
+    pass
+
+def test_Gaussian_random_walk_moment():
+    pass
 
