### 2023-02-23 c6dbe99
Small step size test for HMC
