+++ test/unit/test_nf.py
@@ -1,59 +1,102 @@
-# Import rqSpline model and train it on a multivariate normal distribution
-
 import jax
-import jax.numpy as jnp  # JAX NumPy
-import jax.random as random  # JAX random
-import optax  # Optimizers
-import flax  # Deep Learning library for JAX
-from flax.training import train_state  # Useful dataclass to keep train state
-
-from flowMC.nfmodel.realNVP import RealNVP
+import jax.numpy as jnp
+from flowMC.nfmodel.realNVP import RealNVP, AffineCoupling
 from flowMC.nfmodel.rqSpline import RQSpline
-from flowMC.nfmodel.utils import make_training_loop, sample_nf
 
-def normal(x, mu, sigma):
-    return jnp.exp(-0.5 * ((x - mu) / sigma) ** 2) / (sigma * jnp.sqrt(2 * jnp.pi))
+def test_affine_coupling():
+    n_features = 3
+    n_hidden = 4
+    mask = jnp.array([1, 0, 1])
+    dt = 1
+    x = jnp.array([[1, 2, 3], [4, 5, 6]])
+
+    layer = AffineCoupling(n_features, n_hidden, mask, dt)
+    layerinit = layer.init(jax.random.PRNGKey(0), x)
+    y, log_det = layer.apply({"params": layerinit['params']}, x)
+
+    assert y.shape == x.shape
+    assert log_det.shape == (2,)
+
+    y_inv, log_det_inv = layer.apply({"params": layerinit['params']}, x, method=layer.inverse)
+
+    assert y_inv.shape == x.shape
+    assert log_det_inv.shape == (2,)
+    assert jnp.allclose(x, y_inv)
+    assert jnp.allclose(log_det, -log_det_inv)
 
 
-def test_rqSpline():
-    n_dim = 5
+def test_realnvp():
+    n_features = 3
+    n_hidden = 4
+    n_layer = 2
+    dt = 1
+    x = jnp.array([[1, 2, 3], [4, 5, 6]])
 
-    # Generate data
-    rng_key, rng_data = jax.random.split(jax.random.PRNGKey(0))
-    data = jax.random.multivariate_normal(rng_data, jnp.zeros(n_dim), jnp.eye(n_dim), (10000,))
+    rng_key, rng_subkey = jax.random.split(jax.random.PRNGKey(0), 2)
+    model = RealNVP(n_layer, n_features, n_hidden, dt)
+    model_init = model.init(rng_subkey, jnp.ones((1, n_features)))
+    params = model_init["params"]
+    variables = model_init["variables"]
 
-    key, rng_model, rng_init, rng_train, rng_nf_sample = jax.random.split(
-        jax.random.PRNGKey(0), 5
-    )
+    y, log_det = model.apply({"params": params, "variables": variables}, x)
 
-    # Model parameters
-    n_layers = 8
-    n_hiddens = [64, 64]
+    assert y.shape == x.shape
+    assert log_det.shape == (2,)
+
+    y_inv, log_det_inv = model.apply({"params": params, "variables": variables}, y, method=model.inverse)
+
+    assert y_inv.shape == x.shape
+    assert log_det_inv.shape == (2,)
+    assert jnp.allclose(x, y_inv)
+    assert jnp.allclose(log_det, -log_det_inv)
+
+    rng_key = jax.random.PRNGKey(0)
+    samples = model.apply(
+            {"params": params, "variables": variables},
+            rng_key,
+            2,
+            method=model.sample,
+        )
+
+    assert samples.shape == (2, 3)
+
+    log_prob = model.apply({'params': params, 'variables': variables}, samples, method=model.log_prob)
+
+    assert log_prob.shape == (2,)
+
[…]
