### 2023-02-24 ecdc68d
Add test for MALA and Gaussian
