### 2023-02-24 d28eb7d
Restructure kernel test
