+++ test/unit/test_kernels.py
@@ -8,91 +8,108 @@ import jax.numpy as jnp
 def log_posterior(x):
     return -0.5 * jnp.sum(x ** 2)
 
-def test_HMC():
-    n_dim = 2
-    n_chains = 1
-    HMC_obj = HMC(log_posterior, True, {"step_size": 1,"n_leapfrog": 5, "inverse_metric": jnp.ones(n_dim)})
+class TestHMC:
 
-    rng_key_set = initialize_rng_keys(n_chains, seed=42)
+    def test_HMC_deterministic(self):
+        n_dim = 2
+        n_chains = 1
+        HMC_obj = HMC(log_posterior, True, {"step_size": 1,"n_leapfrog": 5, "inverse_metric": jnp.ones(n_dim)})
 
-    initial_position = jax.random.normal(rng_key_set[0], shape=(n_chains, n_dim)) * 1
-    initial_PE = jax.vmap(HMC_obj.potential)(initial_position)
+        rng_key_set = initialize_rng_keys(n_chains, seed=42)
 
-    HMC_kernel, leapfrog_kernel, leapfrog_step = HMC_obj.make_kernel(return_aux=True)
+        initial_position = jax.random.normal(rng_key_set[0], shape=(n_chains, n_dim)) * 1
+        initial_PE = jax.vmap(HMC_obj.potential)(initial_position)
 
-    # Test whether the HMC kernel is deterministic
+        HMC_kernel, leapfrog_kernel, leapfrog_step = HMC_obj.make_kernel(return_aux=True)
 
-    result1 = (HMC_kernel(rng_key_set[0], initial_position[0], initial_PE[0], HMC_obj.params))
-    result2 = (HMC_kernel(rng_key_set[0], initial_position[0], initial_PE[0], HMC_obj.params))
+        # Test whether the HMC kernel is deterministic
 
-    assert jnp.allclose(result1[0],result2[0])
-    assert result1[1]==result2[1]
-    assert result1[2]==result2[2]
+        result1 = (HMC_kernel(rng_key_set[0], initial_position[0], initial_PE[0], HMC_obj.params))
+        result2 = (HMC_kernel(rng_key_set[0], initial_position[0], initial_PE[0], HMC_obj.params))
 
-    # Test whether the leapfrog kernel is reversible
+        assert jnp.allclose(result1[0],result2[0])
+        assert result1[1]==result2[1]
+        assert result1[2]==result2[2]
 
-    key1, key2 = jax.random.split(rng_key_set[0])
-    initial_momentum = jax.random.normal(key1, shape=initial_position.shape) * jnp.ones(n_dim)**-0.5
+    def test_leapfrog_reversible(self):
+        # Test whether the leapfrog kernel is reversible
+        n_dim = 2
+        n_chains = 1
+        HMC_obj = HMC(log_posterior, True, {"step_size": 1,"n_leapfrog": 5, "inverse_metric": jnp.ones(n_dim)})
 
-    new_position, new_momentum = leapfrog_step(initial_position, initial_momentum, HMC_obj.params)
-    rev_position, rev_momentum = leapfrog_step(new_position, -new_momentum, HMC_obj.params)
+        rng_key_set = initialize_rng_keys(n_chains, seed=42)
+        initial_position = jax.random.normal(rng_key_set[0], shape=(n_chains, n_dim)) * 1
+        initial_PE = jax.vmap(HMC_obj.potential)(initial_position)
 
-    assert jnp.allclose(rev_position, initial_position)
-    assert jnp.allclose(initial_PE, HMC_obj.potential(rev_position))
+        HMC_kernel, leapfrog_kernel, leapfrog_step = HMC_obj.make_kernel(return_aux=True)
[…]
