### 2023-02-24 fe9e04f
Update kernel test and fix rqspline
