+++ test/unit/test_util.py
@@ -1,1 +1,0 @@
-from flowMC.utils.PRNG_keys import initialize_rng_keys
