### 2023-02-28 kazewong/flowMC#76 @marylou-gabrie
Thanks for raising this issue. PR #95 will fix this issue by restructuring tests and removing additional requirements except from `optax ` which is actually itself a requirement installed by `flax`.
