### 2023-02-28 kazewong/flowMC#77 @kazewong
Indeed `pytest` is used to run the tests. We have added a sentence in the requirement section of the README related to this issue in commit #7064998d90bbc96858b70bd4fe43a21f1aeee45a
