### 2023-02-28 3c1697d
Fix typo in rqsplin
