### 2023-02-28 311f04f
Fix HMC docstring inconsistency
