### 2023-02-28 kazewong/flowMC#83 @kazewong
@Daniel-Dodd We have updated the documentation accordingly. We will implement short examples of the objects in a separate issue.
