### 2023-02-28 kazewong/flowMC#85 @kazewong
@Daniel-Dodd We have extended the contribution guide according in #91. We have also added the contribution guide in our documentation. #91
