### 2023-03-02 kazewong/flowMC#79 @matt-graham
> @matt-graham In PR #95, we have implemented a number of unit tests including:
>
> 1. Check whether the Markov kernels are deterministic.
> 2. Check whether a small step-size parameter will ensure acceptance.
> 3. Check whether leapfrog integrator of HMC is reversible.
> 4. Check whether kernels can match empirical moments of a normal distribution.
> 5. Check the normalizing flow are bijections.
>
> We have not implemented this test (train the normalizing flow model to learn a simple target distribution such as a Gaussian) since our based distribution of the normalizing flow is a multivariate normal distribution, and we whiten the data when we train the flow, so there is no training needed to approximate a normal distribution. On the other hand, there is an example showing the normalizing flow can learn to represent a non-trivial distribution such as the dual moon distribution.
>
> Please let us know if you think this test is necessarily for the JOSS paper to proceed. We can also open another issue to incorporate the dual moon example into the test suite, however given the automated test ran on github is on CPU, this would strain the testing pipeline quite a bit, so we prefer to leave it out for the moment

Hi @kazewong. Apologies for being a bit slow getting back to you on this. The unit tests added in #95 looks great to me. What you say about not testing ability of normalizing flow to learn simple (multivariate Gaussian) target distribution makes sense, and agree that not appropriate to run a more complicated example on CPU on GitHub Actions runners.
