### 2023-03-02 kazewong/flowMC#76 @matt-graham
Thanks for fixing.

With regards to `optax`, not a big issue but I would say any module you directly import its worth explicitly including in requirements rather than relying on its being indirectly installed because it is a dependency of another dependency (`flax`), as this guard against the possibility of `optax` being removed as a dependency by `flax` in some future version.
