### 2023-03-02 kazewong/flowMC#100 @matt-graham
Tests fail on Python 3.8 and 3.9 due to upstream issue in `distrax`
Raising issue as part of JOSS review https://github.com/openjournals/joss-reviews/issues/5021

Running the tests in clean Python 3.8 or 3.9 environments I get a series of
```
TypeError: Subscripted generics cannot be used with class and instance checks
```
errors which appear to be due to deepmind/distrax/issues/224. It looks like the most recent 0.1.3 release of `distrax` on PyPI fixes these issues as manually updating to that version and rerunning the tests in the Python 3.8 and Python 3.9 environments fixes the errors, so just bumping the pinned version of `distrax` to 0.1.3 in `setup.cfg` should fix.
