### 2023-03-02 kazewong/flowMC#100 @matt-graham
It may also be worth changing the Python versions tested over in the GitHub Actions workflow
https://github.com/kazewong/flowMC/blob/239c737856c5f3c3bce473f973253580a8cb85c3/.github/workflows/python-package.yml#L19
to include Python 3.8 and 3.9 to automatically catch these sort of problems
