### 2023-03-02 kazewong/flowMC#100 @kazewong
@matt-graham Thanks for the catch. We saw the same problem at some point and I thought I had fixed it by bumping `distrax` to version 0.1.3, but the commit must have been lost somewhere.

Anyway, this is now fixed in #102. The automated tests now runs on python 3.8, 3.9 3.10 and they all passes.
