### 2023-03-03 #5021 @rkurchin
Minor editorial comments:
- line 37: hyphenate "black-box" (compound adjective modifying "interface") and replace "by" with "with"
- 38: I would rephrase to "...default parameters, as well as an extensive guide..."
- 48: add a comma after "kernel"
- 69-70: rephrase to "until chains of the desired length are obtained"
- 83-85: sentence is a bit long and hard to parse, maybe just put in numbers, as in "The only inputs we require from the users are (1) [...], and (2) [...]"

Otherwise, it all looks good! (apart from missing DOI above)
