### 2023-03-07 #5021 @kazewong
@rkurchin The version number is v0.1 https://github.com/kazewong/flowMC/releases/tag/latest and the DOI on zenodo is https://doi.org/10.5281/zenodo.7706605
